+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleBannerAdapter.java
@@ -5,7 +5,6 @@ import static com.google.ads.mediation.vungle.VungleMediationAdapter.ERROR_VUNGL
 
 import android.content.Context;
 import android.util.Log;
-import android.view.View;
 import android.widget.RelativeLayout;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -18,12 +17,11 @@ import com.google.android.gms.ads.AdSize;
 import com.google.android.gms.ads.mediation.MediationBannerAdapter;
 import com.google.android.gms.ads.mediation.MediationBannerListener;
 import com.vungle.warren.AdConfig;
+import com.vungle.warren.BannerAdConfig;
 import com.vungle.warren.Banners;
 import com.vungle.warren.LoadAdCallback;
 import com.vungle.warren.PlayAdCallback;
-import com.vungle.warren.Vungle;
 import com.vungle.warren.VungleBanner;
-import com.vungle.warren.VungleNativeAd;
 import com.vungle.warren.error.VungleException;
 
 public class VungleBannerAdapter implements PlayAdCallback {
@@ -173,11 +171,7 @@ public class VungleBannerAdapter implements PlayAdCallback {
   }
 
   void preCache() {
-    if (AdConfig.AdSize.isBannerAdSize(mAdConfig.getAdSize())) {
-      Banners.loadBanner(placementId, mAdConfig.getAdSize(), null);
-    } else {
-      Vungle.loadAd(placementId, null);
-    }
+    Banners.loadBanner(placementId, new BannerAdConfig(mAdConfig), null);
   }
 
   void updateVisibility(boolean visible) {
@@ -188,9 +182,6 @@ public class VungleBannerAdapter implements PlayAdCallback {
     this.mVisibility = visible;
     if (vungleBannerAd.getVungleBanner() != null) {
       vungleBannerAd.getVungleBanner().setAdVisibility(visible);
-    }
-    if (vungleBannerAd.getVungleMRECBanner() != null) {
-      vungleBannerAd.getVungleMRECBanner().setAdVisibility(visible);
     }
   }
 
@@ -215,11 +206,7 @@ public class VungleBannerAdapter implements PlayAdCallback {
 
   private void loadBanner() {
     Log.d(TAG, "loadBanner: " + this);
-    if (AdConfig.AdSize.isBannerAdSize(mAdConfig.getAdSize())) {
-      Banners.loadBanner(placementId, mAdConfig.getAdSize(), mAdLoadCallback);
-    } else {
-      Vungle.loadAd(placementId, mAdLoadCallback);
-    }
+    Banners.loadBanner(placementId, new BannerAdConfig(mAdConfig), mAdLoadCallback);
   }
 
   private void createBanner() {
@@ -238,7 +225,7 @@ public class VungleBannerAdapter implements PlayAdCallback {
 
     if (AdConfig.AdSize.isBannerAdSize(mAdConfig.getAdSize())) {
       VungleBanner vungleBanner = Banners
-          .getBanner(placementId, mAdConfig.getAdSize(), playAdCallback);
+          .getBanner(placementId, new BannerAdConfig(mAdConfig), playAdCallback);
       if (vungleBanner != null) {
         Log.d(TAG, "display banner:" + vungleBanner.hashCode() + this);
         if (vungleBannerAd != null) {
@@ -261,33 +248,12 @@ public class VungleBannerAdapter implements PlayAdCallback {
         }
       }
     } else {
-      View adView = null;
-      VungleNativeAd vungleMRECBanner = Vungle.getNativeAd(placementId, mAdConfig, playAdCallback);
-      if (vungleMRECBanner != null) {
-        adView = vungleMRECBanner.renderNativeView();
-      }
-
-      if (adView != null) {
-        Log.d(TAG, "display MREC:" + vungleMRECBanner.hashCode() + this);
-        if (vungleBannerAd != null) {
-          vungleBannerAd.setVungleMRECBanner(vungleMRECBanner);
-        }
-
-        updateVisibility(mVisibility);
-        adView.setLayoutParams(adParams);
-        // don't add to parent here
-        if (mediationAdapter != null && mediationListener != null) {
-          mediationListener.onAdLoaded(mediationAdapter);
-        }
-      } else {
-
-        if (mediationAdapter != null && mediationListener != null) {
+      if (mediationAdapter != null && mediationListener != null) {
           AdError error = new AdError(ERROR_VUNGLE_BANNER_NULL,
               "Vungle SDK returned a successful load callback, but Banners.getBanner() or Vungle.getNativeAd() returned null.",
               ERROR_DOMAIN);
           Log.d(TAG, error.getMessage());
-          mediationListener.onAdFailedToLoad(mediationAdapter, error);
-        }
+        mediationListener.onAdFailedToLoad(mediationAdapter, error);
       }
     }
   }
@@ -314,6 +280,11 @@ public class VungleBannerAdapter implements PlayAdCallback {
     if (vungleBannerAd != null) {
       vungleBannerAd.detach();
     }
+  }
+
+  @Override
+  public void creativeId(String creativeId) {
+    // no-op
   }
 
   /**