+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleInterstitialAdapter.java
@@ -1,17 +1,3 @@
-// Copyright 2014 Google Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package com.vungle.mediation;
 
 import static com.google.ads.mediation.vungle.VungleMediationAdapter.ERROR_AD_ALREADY_LOADED;
@@ -121,7 +107,7 @@ public class VungleInterstitialAdapter
   }
 
   private void loadAd() {
-    if (mVungleManager.isAdPlayable(mPlacementForPlay)) {
+    if (Vungle.canPlayAd(mPlacementForPlay)) {
       if (mMediationInterstitialListener != null) {
         mMediationInterstitialListener.onAdLoaded(VungleInterstitialAdapter.this);
       }
@@ -161,6 +147,12 @@ public class VungleInterstitialAdapter
   @Override
   public void showInterstitial() {
     Vungle.playAd(mPlacementForPlay, mAdConfig, new PlayAdCallback() {
+
+      @Override
+      public void creativeId(String creativeId) {
+        // no-op
+      }
+
       @Override
       public void onAdStart(String placementID) {
         if (mMediationInterstitialListener != null) {